product details Edited

+++ routes/admin.js
@@ -29,13 +29,21 @@ router.post('/add-product',(req,res)=>{
  
   })
 })
-router.get('/delete-product/:id',(req,res)=>{
-  let productId=req.params.id
-  console.log(productId)
-  productHelper.deleteProduct(productId).then((response)=>{
-    res.redirect('admin/')
-  })
+const { ObjectId } = require('mongodb');
 
-})
+router.get('/delete-product/:id', (req, res) => {
+  let productName = req.params.id;
+  console.log(productName)
+  if (!ObjectId.isValid(productId)) {
+    console.log('Invalid productId:', productId);
+    return res.redirect('admin/');
+   
+  }
+  productHelper.deleteProduct(productId).then((response) => {
+    res.redirect('admin/');
+ 
+  });
+});
+
 
 module.exports = router;
